Fix: bugs de compilation avec Opus 4 Partie 2 espoir 3

+++ src/main/java/com/regulation/contentieux/model/Affaire.java
@@ -261,7 +261,7 @@ public class Affaire {
         this.montantTotal = montantAmendeTotal; // Synchroniser
     }
 
-    public BigDecimal getMontantAmende() {
+    public long getMontantAmende() {
         return getMontantAmendeTotal(); // Alias pour RapportService
     }
